chasing update

+++ Assets/Scripts/Game Play Subsystem/Game Play Manager/DamageInfo.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 public class DamageInfo
 {
     public GameObject attacker;
-    public float damageValue;
+    public int damageValue;
 }
